+++ demo/admin/src/products/categories/ProductCategoriesTable.tsx
@@ -1,6 +1,8 @@
 import { useQuery } from "@apollo/client";
 import {
     CrudContextMenu,
+    filterByFragment,
+    GridColDef,
     GridFilterButton,
     MainContent,
     muiGridFilterToGql,
@@ -17,8 +19,7 @@ import {
 } from "@comet/admin";
 import { Add as AddIcon, Edit } from "@comet/admin-icons";
 import { Button, IconButton } from "@mui/material";
-import { DataGridPro, GridColDef, GridToolbarQuickFilter } from "@mui/x-data-grid-pro";
-import { filter } from "graphql-anywhere";
+import { DataGridPro, GridToolbarQuickFilter } from "@mui/x-data-grid-pro";
 import gql from "graphql-tag";
 import * as React from "react";
 import { FormattedMessage } from "react-intl";
@@ -90,7 +91,7 @@ const columns: GridColDef<GQLProductsCategoriesListFragment>[] = [
                         }}
                         refetchQueries={["ProductCategoriesList"]}
                         copyData={() => {
-                            return filter<GQLProductsCategoriesListFragment>(productCategoriesFragment, params.row);
+                            return filterByFragment<GQLProductsCategoriesListFragment>(productCategoriesFragment, params.row);
                         }}
                     />
                 </>