+++ demo/admin/src/products/future/ProductsPage.tsx
@@ -1,4 +1,7 @@
 import {
+    MainContent,
+    RouterTab,
+    RouterTabs,
     SaveBoundary,
     SaveBoundarySaveButton,
     Stack,
@@ -10,10 +13,12 @@ import {
     ToolbarBackButton,
     ToolbarFillSpace,
 } from "@comet/admin";
+import { ProductVariantsGrid } from "@src/products/future/generated/ProductVariantsGrid";
 import * as React from "react";
 import { useIntl } from "react-intl";
 
 import { ProductForm } from "./generated/ProductForm";
+import { ProductPriceForm } from "./generated/ProductPriceForm";
 import { ProductsGrid } from "./generated/ProductsGrid";
 
 export function ProductsPage(): React.ReactElement {
@@ -22,7 +27,9 @@ export function ProductsPage(): React.ReactElement {
         <Stack topLevelTitle={intl.formatMessage({ id: "products.products", defaultMessage: "Products" })}>
             <StackSwitch>
                 <StackPage name="grid">
-                    <ProductsGrid />
+                    <MainContent fullHeight disablePadding>
+                        <ProductsGrid />
+                    </MainContent>
                 </StackPage>
                 <StackPage name="edit" title={intl.formatMessage({ id: "products.editProduct", defaultMessage: "Edit Product" })}>
                     {(selectedProductId) => (
@@ -35,8 +42,30 @@ export function ProductsPage(): React.ReactElement {
                                     <SaveBoundarySaveButton />
                                 </ToolbarActions>
                             </StackToolbar>
-                            <ProductForm id={selectedProductId} />
+                            <RouterTabs>
+                                <RouterTab
+                                    forceRender={true}
+                                    path=""
+                                    label={intl.formatMessage({ id: "products.product", defaultMessage: "Product" })}
+                                >
+                                    <ProductForm id={selectedProductId} />
+                                </RouterTab>
+                                <RouterTab
+                                    forceRender={true}
+                                    path="/price"
+                                    label={intl.formatMessage({ id: "products.price", defaultMessage: "Price" })}
+                                >
+                                    <ProductPriceForm id={selectedProductId} />
+                                </RouterTab>
+                            </RouterTabs>
                         </SaveBoundary>
+                    )}
+                </StackPage>
+                <StackPage name="variants" title={intl.formatMessage({ id: "products.editProduct", defaultMessage: "Product variants" })}>
+                    {(selectedId) => (
+                        <MainContent fullHeight disablePadding>
+                            <ProductVariantsGrid product={selectedId} />
+                        </MainContent>
                     )}
                 </StackPage>
                 <StackPage name="add" title={intl.formatMessage({ id: "products.addProduct", defaultMessage: "Add Product" })}>