+++ packages/eslint-config/react.js
@@ -36,13 +36,23 @@ module.exports = {
                     },
                     {
                         name: "@mui/material",
-                        importNames: ["Badge"],
-                        message: "Please use Badge from @comet/admin instead",
-                    },
-                    {
-                        name: "@mui/material",
                         importNames: ["Alert"],
                         message: "Please use Alert from @comet/admin instead",
+                    },
+                    {
+                        name: "@mui/x-data-grid",
+                        importNames: ["GridColDef"],
+                        message: "Please use GridColDef from @comet/admin instead",
+                    },
+                    {
+                        name: "@mui/x-data-grid-pro",
+                        importNames: ["GridColDef"],
+                        message: "Please use GridColDef from @comet/admin instead",
+                    },
+                    {
+                        name: "@mui/x-data-grid-premium",
+                        importNames: ["GridColDef"],
+                        message: "Please use GridColDef from @comet/admin instead",
                     },
                 ],
             },